+++ src/main/java/com/jobnote/domain/user/service/AuthTokenService.java
@@ -3,9 +3,9 @@ package com.jobnote.domain.user.service;
 import com.jobnote.auth.token.Token;
 import com.jobnote.auth.token.TokenClaim;
 import com.jobnote.auth.token.TokenProvider;
-import com.jobnote.domain.user.domain.RefreshToken;
+import com.jobnote.domain.refreshtoken.domain.RefreshToken;
 import com.jobnote.domain.user.domain.User;
-import com.jobnote.domain.user.repository.RefreshTokenRepository;
+import com.jobnote.domain.refreshtoken.repository.RefreshTokenRepository;
 import com.jobnote.global.exception.JobNoteException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
